+++ src/constants/connections/integrationsDataKeys.constants.ts
@@ -1,0 +1,4 @@
+export const integrationDataKeys = {
+	auth0: ["cid", "client_id", "client_secret", "auth0_domain"],
+	google: ["auth_scopes", "cal_id", "form_id"],
+};