+++ src/validations/index.ts
@@ -17,10 +17,13 @@ export {
 	googleCalendarIntegrationSchema,
 	googleFormsIntegrationSchema,
 	asanaIntegrationSchema,
+	auth0IntegrationSchema,
 } from "@validations/connection.schema";
 export { codeAssetsSchema } from "@validations/coseAndAssets.schema";
 export { manualRunSchema } from "@validations/manualRun.schema";
-export { newOrganizationSchema } from "@validations/organization.schema";
+// eslint-disable-next-line @liferay/sort-exports
+export { selectSchema } from "@validations/shared.schema";
+export { newOrganizationSchema, organizationSchema, newOrganizationUserSchema } from "@validations/organization.schema";
 export type { TriggerFormData } from "@validations/trigger.schema";
 export { triggerSchema, triggerResolver } from "@validations/trigger.schema";
 export { newVariableShema } from "@validations/variable.schema";